requirements.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
coverage==5.3
Django==3.1.3
django-filter==2.4.0
django-nose==1.4.7
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
dnspython==1.16.0
eventlet==0.29.1
flake8==3.8.4
greenlet==0.4.17
gunicorn==20.0.4
idna==2.10
Markdown==3.3.3
mccabe==0.6.1
model-bakery==1.2.1
nose==1.3.7
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
python-decouple==3.3
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
